Seperate Utils.h members into individual files
