+++ RayTracing/imgui.ini
@@ -27,13 +27,13 @@ Collapsed=0
 
 [Window][Specs]
 Pos=0,26
-Size=1600,157
+Size=1600,178
 Collapsed=0
 DockId=0x00000005,0
 
 [Window][Control]
-Pos=0,428
-Size=1600,472
+Pos=0,519
+Size=1600,381
 Collapsed=0
 DockId=0x00000004,0
 
@@ -44,17 +44,17 @@ Size=1920,1122
 Collapsed=0
 
 [Window][Preview]
-Pos=0,185
-Size=1600,241
+Pos=0,206
+Size=1600,311
 Collapsed=0
 DockId=0x00000003,0
 
 [Docking][Data]
-DockSpace       ID=0xF2944C7F Window=0x4647B76E Pos=86,135 Size=1600,874 Split=Y
-  DockNode      ID=0x00000005 Parent=0xF2944C7F SizeRef=1262,157 Selected=0x4098A460
-  DockNode      ID=0x00000008 Parent=0xF2944C7F SizeRef=1262,715 Split=X
+DockSpace       ID=0xF2944C7F Window=0x4647B76E Pos=138,187 Size=1600,874 Split=Y
+  DockNode      ID=0x00000005 Parent=0xF2944C7F SizeRef=1262,178 Selected=0x4098A460
+  DockNode      ID=0x00000008 Parent=0xF2944C7F SizeRef=1262,803 Split=X
     DockNode    ID=0x00000001 Parent=0x00000008 SizeRef=497,874 Selected=0x9B88CA06
     DockNode    ID=0x00000002 Parent=0x00000008 SizeRef=1101,874 Split=Y Selected=0xCE6F6A26
-      DockNode  ID=0x00000003 Parent=0x00000002 SizeRef=1600,241 Selected=0x382916D5
-      DockNode  ID=0x00000004 Parent=0x00000002 SizeRef=1600,472 CentralNode=1 Selected=0xCE6F6A26
+      DockNode  ID=0x00000003 Parent=0x00000002 SizeRef=1600,311 Selected=0x382916D5
+      DockNode  ID=0x00000004 Parent=0x00000002 SizeRef=1600,490 CentralNode=1 Selected=0xCE6F6A26
