+++ RayTracing/src/Core/Material/Metal.cpp
@@ -1,7 +1,8 @@
 #include "Metal.h"
 
 #include "Core/Object/HittableObject.h"
-#include "Core/Utils.h"
+#include "Utils/Utils.h"
+#include "Utils/Random.h"
 #include "Core/Texture/SolidColorTexture.h"
 
 Metal::Metal(Color& color, float fuzz)
