+++ RayTracing/src/Core/Object/RotateY.cpp
@@ -1,6 +1,6 @@
 #include "RotateY.h"
 
-#include "Core/Utils.h"
+#include "Utils/Utils.h"
 
 RotateY::RotateY(std::shared_ptr<HittableObject>& object, float angle)
 	: m_Object(object)
