+++ RayTracing/src/Utils/Color.h
@@ -1,18 +1,10 @@
 #pragma once
 
-#include "AABB.h"
-#include "Ray.h"
-
-#include <cstdlib>
-
-#include "Walnut/Random.h"
 #include "glm/glm.hpp"
 
-#include <ctime>
-#include <limits>
+#include <cstdint>
 
 namespace Utils {
-
 	namespace Color {
 		enum {
 			KeepColorChannel = 0xCC5
@@ -196,175 +188,4 @@ namespace Utils {
 			return Vec4ToRGBA(Vec4ToRGBABlendColor(foregroundVec, backgroundVec));
 		}
 	}
-
-	namespace Time {
-
-		struct TimeComponents {
-			uint64_t milli_seconds;
-			uint64_t seconds;
-			uint64_t minutes;
-			uint64_t hours;
-			uint64_t days;
-			uint64_t months;
-			uint64_t years;
-			std::time_t time;
-		};
-
-		static TimeComponents GetTime(std::time_t time)
-		{
-			std::tm* st = std::localtime(&time);
-			return { time % 1000ULL,
-					 time /    1000ULL % 60ULL,
-					 time /   60000ULL % 60ULL,
-					 time / 3600000ULL % 24ULL,
-					 (uint64_t) st->tm_mday,
-					 (uint64_t) st->tm_mon,
-					 (uint64_t) st->tm_year,
-					 time};
-		}
-
-		static void GetTime(TimeComponents& components, std::time_t time)
-		{
-			std::tm* st = std::localtime(&time);
-			components.milli_seconds = time % 1000ULL;
-			components.seconds       = time /    1000ULL % 60ULL;
-			components.minutes       = time /   60000ULL % 60ULL;
-			components.hours         = time / 3600000ULL % 24ULL;
-			components.days			 = (uint64_t)st->tm_mday;
-			components.months		 = (uint64_t)st->tm_mon;
-			components.years		 = (uint64_t)st->tm_year;
-			components.time			 = time;
-		}
-
-	}
-
-	class Random
-	{
-	public:
-
-		static inline void Init()
-		{
-			s_Random.Init();
-		}
-
-		static inline double RandomDouble()
-		{
-			// Returns a random real in [0,1).
-			//float a = s_Random.Float();
-			return s_Random.Float();// rand() / (RAND_MAX + 1.0f);
-		}
-
-		static inline double RandomDouble(double min, double max)
-		{
-			// Returns a random real in [min,max).
-			return min + (max - min) * RandomDouble();
-		}
-
-		static inline int32_t RandomInt()
-		{
-			return static_cast<int32_t>(RandomDouble());
-		}
-
-		static inline int32_t RandomInt(int32_t min, int32_t max)
-		{
-			return static_cast<int32_t>(RandomDouble(min, max));
-		}
-
-		static inline float RandomFloat()
-		{
-			return static_cast<float>(RandomDouble());
-		}
-
-		static inline float RandomFloat(float min, float max)
-		{
-			return static_cast<float>(RandomDouble(min, max));
-		}
-
-		template<typename T>
-		static inline T GetRandom()
-		{
-			return static_cast<T>(RandomDouble());
-		}
-
-		template<typename T>
-		static inline T GetRandom(T min, T max)
-		{
-			return static_cast<T>(RandomDouble(min, max));
-		}
-
-		static inline glm::vec3 RandomVec3()
-		{
-			return { RandomDouble(), RandomDouble(), RandomDouble() };
-		}
-
-		static inline glm::vec3 RandomVec3(double min, double max)
-		{
-			return { RandomDouble(min, max), RandomDouble(min,  max), RandomDouble(min, max) };
-		}
-
-		static inline glm::vec3 RandomInUnitSphere()
-		{
-			//return Random::RandomVec3(-1.0, 1.0);
-			while (true)
-			{
-				glm::vec3 p = Random::RandomVec3(-1.0, 1.0);
-				if (glm::dot(p, p) >= 1.0f)
-					continue;
-				return p;
-			}
-		}
-
-		static inline glm::vec3 RandomInUnitDisk()
-		{
-			//return Random::RandomVec3(-1.0, 1.0);
-			while (true)
-			{
-				glm::vec3 p = glm::vec3(Random::RandomDouble(-1.0, 1.0), Random::RandomDouble(-1.0, 1.0), 0.0);
-				if (glm::dot(p, p) >= 1.0f)
-					continue;
-				return p;
-			}
-		}
-
-		static inline glm::vec3 RandomUnitVector()
-		{
-			glm::vec3 tmp = RandomInUnitSphere();
-			return tmp / glm::length(tmp);
-		}
-
-		static inline glm::vec3 RandomInHemisphere(Ray::Vec3& normal)
-		{
-			glm::vec3 in_unit_sphere = RandomInUnitSphere();
-			if (glm::dot(in_unit_sphere, normal) > 0.0f)
-			{
-				return in_unit_sphere;
-			}
-			else
-			{
-				return -in_unit_sphere;
-			}
-		}
-
-	private:
-
-		static Walnut::Random s_Random;
-
-	};
-
-	static AABB SurroundingBox(const AABB& box0, const AABB& box1)
-	{
-		glm::vec3 small(glm::min(box0.GetMin(), box1.GetMin()));
-		glm::vec3 big(glm::max(box0.GetMax(), box1.GetMax()));
-
-		return AABB(small, big);
-	}
-
-	template<typename T>
-	static T UnitVec(T t)
-	{
-		return t / glm::length(t);
-	}
-
-	constexpr static float infinity = std::numeric_limits<float>::infinity();
-
 }
